resttructure FirebaseAuthManager
