+++ App/CourseApp/Networking/Models/Credentials.swift
@@ -1,0 +1,11 @@
+//
+//  Credentials.swift
+//  CourseApp
+//
+//  Created by Marcel Mravec on 18.06.2024.
+//
+
+struct Credentials {
+    let email: String
+    let password: String
+}
